--- FluentAvalonia/UI/Controls/InfoBar/InfoBarEventArgs.cs
@@ -1,25 +1,0 @@
-﻿using System;
-
-namespace FluentAvalonia.UI.Controls
-{
-	public class InfoBarClosedEventArgs : EventArgs
-	{
-		internal InfoBarClosedEventArgs(InfoBarCloseReason reason)
-		{
-			Reason = reason;
-		}
-
-		public InfoBarCloseReason Reason { get; }
-	}
-
-	public class InfoBarClosingEventArgs
-	{
-		internal InfoBarClosingEventArgs(InfoBarCloseReason reason)
-		{
-			Reason = reason;
-		}
-
-		public InfoBarCloseReason Reason { get; }
-		public bool Cancel { get; set; }
-	}
-}
